src/main/java/com/yzz/service/impl/ProductCategoryServiceImpl.java
package com.yzz.service.impl;

import com.yzz.dao.ProductCategoryDao;
import com.yzz.domain.ProductCategory;
import com.yzz.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {

    @Autowired
    private ProductCategoryDao productCategoryDao;

    @Override
    public ProductCategory findOne(Integer productCategoryId) {
        return productCategoryDao.findOne(productCategoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryDao.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> productTypeList) {
        return productCategoryDao.findByCategoryTypeIn(productTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryDao.save(productCategory);
    }
}
